chore: convert ohlcAggregator module to TypeScript
